backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.0.3
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
mysql-connector-python==9.4.0
python-dotenv==1.0.1
PyJWT==2.8.0
werkzeug==3.1.5
zipp==3.23.0
pytest==7.4.4
pytest-mock==3.12.0
